+++ core/src/elements/ons-alert-dialog/index.js
@@ -25,7 +25,7 @@ import deviceBackButtonDispatcher from 'ons/device-back-button-dispatcher';
 import DoorLock from 'ons/doorlock';
 
 const scheme = {
-  '': 'alert-dialog--*',
+  '.alert-dialog': 'alert-dialog--*',
   '.alert-dialog-container': 'alert-dialog-container--*',
   '.alert-dialog-title': 'alert-dialog-title--*',
   '.alert-dialog-content': 'alert-dialog-content--*',
@@ -33,8 +33,18 @@ const scheme = {
   '.alert-dialog-button': 'alert-dialog-button--*',
   '.alert-dialog-footer--one': 'alert-dialog-footer--one--*',
   '.alert-dialog-button--one': 'alert-dialog-button--one--*',
-  '.alert-dialog-button--primal': 'alert-dialog-button--primal--*'
+  '.alert-dialog-button--primal': 'alert-dialog-button--primal--*',
+  '.alert-dialog-mask': 'alert-dialog-mask--*'
 };
+
+const templateSource = util.createElement(`
+  <div>
+    <div class="alert-dialog-mask"></div>
+    <div class="alert-dialog">
+      <div class="alert-dialog-container"></div>
+    </div>
+  </div>
+`);
 
 const _animatorDict = {
   'default': platform.isAndroid() ? AndroidAlertDialogAnimator : IOSAlertDialogAnimator,
@@ -188,19 +198,46 @@ class AlertDialogElement extends BaseElement {
     return util.findChild(this.children[0], '.alert-dialog-title');
   }
 
-  get _contentElement() {
-    return util.findChild(this.children[0], '.alert-dialog-content');
+  /**
+   * @return {Element}
+   */
+  get _mask() {
+    return util.findChild(this, '.alert-dialog-mask');
   }
 
+  /**
+   * @return {Element}
+   */
   get _dialog() {
-    return this;
+    return util.findChild(this, '.alert-dialog');
+
+  }
+
+  /**
+   * @return {Element}
+   */
+  get _titleElement() {
+    return util.findChild(this._dialog.children[0], '.alert-dialog-title');
+  }
+
+  /**
+   * @return {Element}
+   */
+  get _contentElement() {
+    return util.findChild(this._dialog.children[0], '.alert-dialog-content');
   }
 
   createdCallback() {
-    this._compile();
-    this._mask = this._createMask(this.getAttribute('mask-color'));
+    if (!this.hasAttribute('_compiled')) {
+      this._compile();
+      ModifierUtil.initModifier(this, scheme);
 
-    ModifierUtil.initModifier(this, scheme);
+      this.setAttribute('_compiled', '');
+    }
+
+    this._visible = false;
+    this._doorLock = new DoorLock();
+    this._boundCancel = this._cancel.bind(this);
 
     this._animatorFactory = new AnimatorFactory({
       animators: _animatorDict,
@@ -208,16 +245,34 @@ class AlertDialogElement extends BaseElement {
       baseClassName: 'AlertDialogAnimator',
       defaultAnimation: this.getAttribute('animation')
     });
-
-    this._visible = false;
-    this._doorLock = new DoorLock();
-    this._boundCancel = this._cancel.bind(this);
   }
 
   _compile() {
+    const style = this.getAttribute('style');
+
     this.style.display = 'none';
-    this.style.zIndex = '20001';
-    this.classList.add('alert-dialog');
+
+    const template = templateSource.cloneNode(true);
+    const alertDialog = template.children[1];
+
+    if (style) {
+      alertDialog.setAttribute('style', style);
+    }
+
+    while (this.firstChild) {
+      alertDialog.children[0].appendChild(this.firstChild);
+    }
+
+    while (template.firstChild) {
+      this.appendChild(template.firstChild);
+    }
+
+    this._dialog.style.zIndex = 20001;
+    this._mask.style.zIndex = 20000;
+
+    if (this.getAttribute('mask-color')) {
+      this._mask.style.backgroundColor = this.getAttribute('mask-color');
+    }
   }
 
   /**
@@ -405,10 +460,6 @@ class AlertDialogElement extends BaseElement {
     if (this.parentElement) {
       this.parentElement.removeChild(this);
     }
-
-    if (this._mask.parentElement) {
-      this._mask.parentElement.removeChild(this._mask);
-    }
   }
 
   /**
@@ -441,20 +492,6 @@ class AlertDialogElement extends BaseElement {
         }
       });
     }
-  }
-
-  _createMask(color) {
-    this._mask = util.createElement('<div></div>');
-    this._mask.classList.add('alert-dialog-mask');
-    this._mask.style.zIndex = 20000;
-    this._mask.style.display = 'none';
-
-    if (color) {
-      this._mask.style.backgroundColor = color;
-    }
-
-    document.body.appendChild(this._mask);
-    return this._mask;
   }
 
   attachedCallback() {