+++ core/src/elements/ons-list.spec.js
@@ -25,4 +25,20 @@ describe('ons-list', () => {
     expect(element.classList.contains('list--piyo')).to.be.true;
     expect(element.classList.contains('list--fuga')).to.be.true;
   });
+
+  describe('#_compile()', () => {
+    it('does not compile twice', () => {
+      let div1 = document.createElement('div');
+      let div2 = document.createElement('div');
+      div1.innerHTML = `
+        <ons-list>
+          <ons-list-header>Content</ons-list-header>
+          <ons-list-item>Content</ons-list-item>
+          <ons-list-item>Content</ons-list-item>
+        </ons-list>
+      `;
+      div2.innerHTML = div1.innerHTML;
+      expect(div1.isEqualNode(div2)).to.be.true;
+    });
+  });
 });