+++ core/src/elements/ons-navigator/index.spec.js
@@ -545,4 +545,14 @@ describe('OnsNavigatorElement', () => {
       window.OnsNavigatorElement.registerAnimator('hoge', MyAnimator);
     });
   });
+
+  describe('#_compile()', () => {
+    it('does not compile twice', () => {
+      let div1 = document.createElement('div');
+      let div2 = document.createElement('div');
+      div1.innerHTML = '<ons-navigator></ons-navigator>';
+      div2.innerHTML = div1.innerHTML;
+      expect(div1.isEqualNode(div2)).to.be.true;
+    });
+  });
 });