+++ core/src/elements/ons-progress-circular.js
@@ -79,9 +79,12 @@ class ProgressCircularElement extends BaseElement {
    */
 
   createdCallback() {
-    this._compile();
+    if (!this.hasAttribute('_compiled')) {
+      this._compile();
+      ModifierUtil.initModifier(this, scheme);
 
-    ModifierUtil.initModifier(this, scheme);
+      this.setAttribute('_compiled', '');
+    }
   }
 
   attributeChangedCallback(name, last, current) {