+++ core/src/elements/ons-splitter/index.spec.js
@@ -68,7 +68,7 @@ describe('OnsSplitterElement', () => {
     it('should open right ons-splitter-side', () => {
       expect(splitter.openRight()).to.be.true;
       expect(splitter.openLeft()).to.be.false;
-      expect(splitter.rightIsOpened()).to.be.true;
+      expect(splitter.rightIsOpen()).to.be.true;
     });
   });
 
@@ -80,12 +80,12 @@ describe('OnsSplitterElement', () => {
 
   describe('#closeRight()', () => {
     it('should close right ons-splitter-side', (done) => {
-      expect(splitter.rightIsOpened()).to.be.false;
+      expect(splitter.rightIsOpen()).to.be.false;
 
       expect(splitter.openRight({callback: () => {
-        expect(splitter.rightIsOpened()).to.be.true;
+        expect(splitter.rightIsOpen()).to.be.true;
         expect(splitter.closeRight({callback: () => {
-          expect(splitter.rightIsOpened()).to.be.false;
+          expect(splitter.rightIsOpen()).to.be.false;
           done();
         }})).to.be.true;
       }})).to.be.true;
@@ -101,6 +101,22 @@ describe('OnsSplitterElement', () => {
   describe('#getDeviceBackButtonHandler()', () => {
     it('should return handler object', () => {
       expect(splitter.getDeviceBackButtonHandler()).to.be.an('object');
+    });
+  });
+
+  describe('#_compile()', () => {
+    it('does not compile twice', () => {
+      let div1 = document.createElement('div');
+      let div2 = document.createElement('div');
+      div1.innerHTML = `
+        <ons-splitter>
+          <ons-splitter-side side="left">Left</ons-splitter-side>
+          <ons-splitter-side side="right" collapse>Right</ons-splitter-side>
+          <ons-splitter-content>Content</ons-splitter-content>
+        </ons-splitter>
+      `;
+      div2.innerHTML = div1.innerHTML;
+      expect(div1.isEqualNode(div2)).to.be.true;
     });
   });
 });