+++ core/src/elements/ons-switch.js
@@ -178,8 +178,12 @@ class SwitchElement extends ExtendableLabelElement {
   }
 
   createdCallback() {
-    this._compile();
-    ModifierUtil.initModifier(this, scheme);
+    if (!this.hasAttribute('_compiled')) {
+      this._compile();
+      ModifierUtil.initModifier(this, scheme);
+
+      this.setAttribute('_compiled', '');
+    }
 
     this._updateForCheckedAttribute();
     this._updateForDisabledAttribute();