+++ core/src/elements/ons-tabbar/index.js
@@ -204,16 +204,21 @@ class TabbarElement extends BaseElement {
   createdCallback() {
     this._tabbarId = generateId();
 
+    if (!this.hasAttribute('_compiled')) {
+      this._compile();
+      ModifierUtil.initModifier(this, scheme);
+
+      this.setAttribute('_compiled', '');
+    }
+
+    this._contentElement = util.findChild(this, '.tab-bar__content');
+
     this._animatorFactory = new AnimatorFactory({
       animators: _animatorDict,
       baseClass: TabbarAnimator,
       baseClassName: 'TabbarAnimator',
       defaultAnimation: this.getAttribute('animation')
     });
-
-    this._compile();
-    this._contentElement = util.findChild(this, '.tab-bar__content');
-    ModifierUtil.initModifier(this, scheme);
   }
 
   _compile() {
@@ -375,6 +380,11 @@ class TabbarElement extends BaseElement {
       this._oldPageElement = element;
       var animator = this._animatorFactory.newAnimator(options);
 
+
+      if (oldPageElement !== internal.nullElement) {
+        oldPageElement._hide();
+      }
+
       animator.apply(element, oldPageElement, options.selectedTabIndex, options.previousTabIndex, function() {
         if (oldPageElement !== internal.nullElement) {
           if (options._removeElement) {
@@ -383,7 +393,6 @@ class TabbarElement extends BaseElement {
             });
           } else {
             oldPageElement.style.display = 'none';
-            oldPageElement._hide();
           }
         }