+++ core/src/elements/ons-toolbar.js
@@ -75,8 +75,12 @@ class ToolbarElement extends BaseElement {
    */
 
   createdCallback() {
-    this._compile();
-    ModifierUtil.initModifier(this, scheme);
+    if (!this.hasAttribute('_compiled')) {
+      this._compile();
+      ModifierUtil.initModifier(this, scheme);
+
+      this.setAttribute('_compiled', '');
+    }
 
     this._tryToEnsureNodePosition();
     setImmediate(() => this._tryToEnsureNodePosition());