+++ core/src/elements/ons-toolbar.spec.js
@@ -99,5 +99,13 @@ describe('OnsToolbarElement', () => {
       expect(element.children[1].classList.contains('navigation-bar__center')).to.be.true;
       expect(element.children[2].classList.contains('navigation-bar__right')).to.be.true;
     });
+
+    it('does not compile twice', () => {
+      let div1 = document.createElement('div');
+      let div2 = document.createElement('div');
+      div1.innerHTML = '<ons-toolbar><div class="center">Title</div></ons-toolbar>';
+      div2.innerHTML = div1.innerHTML;
+      expect(div1.isEqualNode(div2)).to.be.true;
+    });
   });
 });